% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_list}
\alias{append_list}
\title{append list}
\usage{
append_list(list_obj = list(), ...)
}
\arguments{
\item{list_obj}{list object}

\item{...}{anything to be appended to the list}
}
\value{

}
\description{
append list
}
\examples{
a <- list()
append_list(a, 1, 2, 3)
}
