% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxgen_comment.R
\name{convert_roxygen}
\alias{convert_roxygen}
\title{Convert selected text to roxygen comment}
\usage{
convert_roxygen(txt)
}
\arguments{
\item{txt}{, a string vector}
}
\value{
a string vector with roxygen commet at each prefix
}
\description{
Convert selected text to roxygen comment
}
\examples{
txt <- "a
b
c
d"
convert_roxygen(txt)
}
