% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_css.R
\name{create_css}
\alias{create_css}
\title{Create a css file in the current directory}
\usage{
create_css()
}
\arguments{
\item{css_content}{content of css file.}
}
\value{
Null. Create a .css file in the current directory
}
\description{
Create a css file in the current directory
}
\examples{
create_css()
}
