% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_file.R
\name{current_fname}
\alias{current_fname}
\title{Extract current active file name}
\usage{
current_fname(line)
}
\arguments{
\item{line}{1L-character vector contains the path of current file. Can be passed from `rstudioapi::getActiveDocumentContext()$path`}
}
\value{
the name of current active file
}
\description{
Extract current active file name
}
\examples{
current_fname("~/Work/cheat_sheet/snippets.R")
}
