% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{data_transit}
\alias{data_transit}
\title{data_transit, track data transition from readin files}
\usage{
data_transit(line, target)
}
\arguments{
\item{line}{char, the text line to be read}

\item{target}{char, the key word target}
}
\value{
list of transit and target
}
\description{
data_transit, track data transition from readin files
}
