% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_input.R
\name{extr_input_file}
\alias{extr_input_file}
\title{Extract input data file}
\usage{
extr_input_file(line)
}
\arguments{
\item{line}{a 1L character vector may or may not contains input data line.}
}
\value{
input file name
}
\description{
Extract input data file
}
\examples{
extr_input_file('read.csv("something.csv", header = T)')
}
