% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_output.R
\name{extr_output_file}
\alias{extr_output_file}
\title{Extract the output files from a text line}
\usage{
extr_output_file(line)
}
\arguments{
\item{line}{1L character the text line to be read}
}
\value{
the file name output from the R script
}
\description{
Extract the output files from a text line
}
\examples{
extr_output_file("saveRDS('test.Rds')")
}
