% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_source.R
\name{extr_src_file}
\alias{extr_src_file}
\title{Extract srource file}
\usage{
extr_src_file(line)
}
\arguments{
\item{line}{a 1L character vector may or may not contains source file path.}
}
\value{
source file path/name
}
\description{
Extract srource file
}
\examples{
extr_src_file('source("src/utils/utils.R")')
}
