% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lib.R
\name{extract_lib}
\alias{extract_lib}
\title{Extract library from one line text}
\usage{
extract_lib(line)
}
\arguments{
\item{line}{character text from code}
}
\value{
NULL or library name extracted from text
}
\description{
Extract library from one line text
}
\examples{
extract_lib("library(apple)")
}
