% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_input.R
\name{extract_uniq_inputs}
\alias{extract_uniq_inputs}
\title{Scan R script and extract non-duplicated input data files}
\usage{
extract_uniq_inputs(doc_contents)
}
\arguments{
\item{doc_contents}{a character vector of a R script}
}
\value{
extracted non-duplicated input data files
}
\description{
Scan R script and extract non-duplicated input data files
}
