% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lib.R
\name{extract_uniq_libs}
\alias{extract_uniq_libs}
\title{Extract unique libraries from the R script}
\usage{
extract_uniq_libs(doc_contents)
}
\arguments{
\item{doc_contents}{character vector of R script can be
extracted from rstudioapi::getActiveDocumentContext
function}
}
\value{
unique libraries used in the document
}
\description{
Extract unique libraries from the R script
}
