% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_output.R
\name{extract_uniq_outputs}
\alias{extract_uniq_outputs}
\title{Extract output files from R script}
\usage{
extract_uniq_outputs(doc_contents)
}
\arguments{
\item{doc_contents}{character vector of R script}
}
\value{
a vector of output files written in the R script if any
NULL if there isn't any
}
\description{
Extract output files from R script
}
