% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_source.R
\name{extract_uniq_src}
\alias{extract_uniq_src}
\title{Scan R script and extract non-duplicated source code files}
\usage{
extract_uniq_src(doc_contents)
}
\arguments{
\item{doc_contents}{a character vector of a R script}
}
\value{
extracted non-duplicated source code files
}
\description{
Scan R script and extract non-duplicated source code files
}
