% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author.R
\name{get_author}
\alias{get_author}
\title{get author ID from  working path}
\usage{
get_author(path)
}
\arguments{
\item{path}{a string vector of one indicating current path.}
}
\value{
string, author ID
}
\description{
get author ID from  working path
}
