% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_io.R
\name{get_io}
\alias{get_io}
\title{get_io get input output information.
create S3 class object of io}
\usage{
get_io(filename, ...)
}
\arguments{
\item{filename}{char pathway and R or Rmd file name to be scanned}

\item{...}{other things pass to the function}
}
\value{
S3 class object of io and inheritated from list class
}
\description{
get_io get input output information.
create S3 class object of io
}
