% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{get_names closure function}
\usage{
get_names(pattern_split, pattern)
}
\arguments{
\item{pattern_split}{character, pattern for spliting}

\item{pattern}{character for extract contents}
}
\value{
functions
}
\description{
get_names closure function
}
