% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_triplet.R
\name{get_triplet}
\alias{get_triplet}
\title{get the triplet information and}
\usage{
get_triplet(path)
}
\arguments{
\item{path}{string character of current file path.}
}
\value{
if detected the triplet will return the triplet info. if not, will
return empty space.
}
\description{
FUNCTION_DESCRIPTION
}
