% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_header.R
\name{insert_header}
\alias{insert_header}
\title{Insert the header of R script which contains the libraries used in the script}
\usage{
insert_header()
}
\value{
text summary of R script at the top of the document
}
\description{
Insert the header of R script which contains the libraries used in the script
}
