% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_features.R
\name{md_tab}
\alias{md_tab}
\title{Transform code to markdown table
cells must be separated by colons `:`}
\usage{
md_tab()
}
\value{
markdown table format
}
\description{
Transform code to markdown table
cells must be separated by colons `:`
}
