% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxgen_comment.R
\name{\%>\%}
\alias{\%>\%}
\alias{redo_roxycomment}
\title{Undo roxygen comment}
\usage{
redo_roxycomment(text)
}
\arguments{
\item{text}{a string vector with roxygen comment sign.}
}
\value{
a string vector without roxygen comment
}
\description{
Undo roxygen comment
}
