% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_io.R
\name{print.io}
\alias{print.io}
\title{print input output table}
\usage{
\method{print}{io}(rst)
}
\arguments{
\item{rst}{S3 io class object}
}
\value{
IO summary table
}
\description{
print input output table
}
