% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dup_libs.R
\name{remove_dup_libs}
\alias{remove_dup_libs}
\title{Remove duplicated libraries and insert cleaned up library}
\usage{
remove_dup_libs()
}
\value{
Insert the libraries in the script at the
cursor position
}
\description{
Remove duplicated libraries and insert cleaned up library
}
