% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_quote.R
\name{rm_quotes}
\alias{rm_quotes}
\title{Remove Quotes from Strings}
\usage{
rm_quotes()
}
\value{
A character vector with all quotation marks removed.
}
\description{
This function removes all quotation marks from the input strings.
}
\examples{
rm_quotes("Hello, \"world\"!")
# [1] "Hello, world!"
}
