% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxgen_comment.R
\name{roxygen_quote}
\alias{roxygen_quote}
\title{add roxygen comment with quotation sign.}
\usage{
roxygen_quote(txt)
}
\arguments{
\item{txt}{a string vector.}
}
\value{
quoted roxygen comments
}
\description{
add roxygen comment with quotation sign.
}
\examples{
# ADD_EXAMPLES_HERE
}
