% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxgen_comment.R
\name{selection2roxygen}
\alias{selection2roxygen}
\title{Roxygen comment on currently selected section}
\usage{
selection2roxygen()
}
\value{

}
\description{
Roxygen comment on currently selected section
}
