% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxgen_comment.R
\name{selection2roxygen_quote}
\alias{selection2roxygen_quote}
\title{Transform current text selection into roxygen comment with quotation.}
\usage{
selection2roxygen_quote()
}
\value{
quoted roxygen text
}
\description{
Transform current text selection into roxygen comment with quotation.
}
