% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection2str.R
\name{selection2str}
\alias{selection2str}
\title{make the selected text into string/characters}
\usage{
selection2str()
}
\value{
quoted text
}
\description{
make the selected text into string/characters
}
