% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_triplet_env.R
\name{source_triplet_env}
\alias{source_triplet_env}
\title{Source Triplet Environment}
\usage{
source_triplet_env()
}
\value{
None
}
\description{
This function sources a specific R script located within a triplet directory structure.
It retrieves the active document's path, constructs the triplet path, and sources the
specified script from the constructed path.
}
\examples{
source_triplet_env()
}
