% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer.R
\name{tracer}
\alias{tracer}
\title{trace to trace certain pattern}
\usage{
tracer(get_target, extract_file)
}
\arguments{
\item{get_target}{function object}

\item{extract_file}{function object}
}
\value{

}
\description{
trace to trace certain pattern
}
