% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxgen_comment.R
\name{uncomment_roxygen}
\alias{uncomment_roxygen}
\title{uncomment roxygen format comment}
\usage{
uncomment_roxygen()
}
\value{
text without roxygen format
}
\description{
uncomment roxygen format comment
}
