% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_features.R
\name{padding}
\alias{padding}
\title{Pad the begining and end of a vector. If one wants to evaluate
a text string, use `<< >>` to quote the code to be evaluted.}
\usage{
padding(txt, pattern)
}
\arguments{
\item{txt}{string vector with length of one.}

\item{pattern}{pattern to pad.}
}
\value{
padded string
}
\description{
Pad the begining and end of a vector. If one wants to evaluate
a text string, use `<< >>` to quote the code to be evaluted.
}
